working

+++ vocal_analyzer/feature_extractor.py
@@ -13,16 +13,16 @@ def extract_features(audio_file):
     pitches = pitches[magnitudes > np.median(magnitudes)]
     pitches = pitches[pitches > 0]
     if len(pitches) > 0:
-        min_pitch = np.min(pitches)
-        max_pitch = np.max(pitches)
-        mean_pitch = np.mean(pitches)
+        min_pitch = float(np.min(pitches))
+        max_pitch = float(np.max(pitches))
+        mean_pitch = float(np.mean(pitches))
     else:
-        min_pitch = max_pitch = mean_pitch = 0
+        min_pitch = max_pitch = mean_pitch = 0.0
     # Detect screaming: simple amplitude threshold
     max_amp = np.max(np.abs(y))
     screaming = max_amp > 0.8  # Arbitrary threshold
     return {
-        "tempo": tempo,
+        "tempo": float(tempo),
         "min_pitch": min_pitch,
         "max_pitch": max_pitch,
         "mean_pitch": mean_pitch,
